#  kallsyms: introduce 'reader', an efficient /proc/kallsyms line parser