--- a/internal/kallsyms/kallsyms.go
+++ b/internal/kallsyms/kallsyms.go
@@ -1,14 +1,17 @@
 package kallsyms
 
 import (
-	"bufio"
-	"bytes"
 	"errors"
 	"fmt"
 	"io"
 	"os"
+	"slices"
+	"strconv"
+	"strings"
 	"sync"
 )
+
+var errAmbiguousKsym = errors.New("multiple kernel symbols with the same name")
 
 var kernelModules struct {
 	sync.RWMutex
@@ -16,8 +19,8 @@ var kernelModules struct {
 	kmods map[string]string
 }
 
-// KernelModule returns the kernel module, if any, a probe-able function is contained in.
-func KernelModule(fn string) (string, error) {
+// SymbolModule returns the kernel module providing a given symbol, if any.
+func SymbolModule(name string) (string, error) {
 	kernelModules.RLock()
 	kmods := kernelModules.kmods
 	kernelModules.RUnlock()
@@ -29,7 +32,7 @@ func KernelModule(fn string) (string, error) {
 	}
 
 	if kmods != nil {
-		return kmods[fn], nil
+		return kmods[name], nil
 	}
 
 	f, err := os.Open("/proc/kallsyms")
@@ -37,13 +40,14 @@ func KernelModule(fn string) (string, error) {
 		return "", err
 	}
 	defer f.Close()
-	kmods, err = loadKernelModuleMapping(f)
+
+	kmods, err = symbolKmods(f)
 	if err != nil {
 		return "", err
 	}
 
 	kernelModules.kmods = kmods
-	return kmods[fn], nil
+	return kmods[name], nil
 }
 
 // FlushKernelModuleCache removes any cached information about function to kernel module mapping.
@@ -54,30 +58,42 @@ func FlushKernelModuleCache() {
 	kernelModules.kmods = nil
 }
 
-var errKsymIsAmbiguous = errors.New("ksym is ambiguous")
-
-func loadKernelModuleMapping(f io.Reader) (map[string]string, error) {
+// symbolKmods parses f into a map of function names to kernel module names.
+// Only function symbols (tT) provided by a kernel module are included in the
+// output.
+func symbolKmods(f io.Reader) (map[string]string, error) {
 	mods := make(map[string]string)
-	scanner := bufio.NewScanner(f)
-	for scanner.Scan() {
-		fields := bytes.Fields(scanner.Bytes())
-		if len(fields) < 4 {
+	r := newReader(f)
+	for r.Line() {
+		s, err, skip := parseSymbol(r, []rune{'t', 'T'})
+		if err != nil {
+			return nil, fmt.Errorf("parsing kallsyms line: %w", err)
+		}
+		if skip {
 			continue
 		}
-		switch string(fields[1]) {
-		case "t", "T":
-			mods[string(fields[2])] = string(bytes.Trim(fields[3], "[]"))
-		default:
+
+		// Lines without a module will have an empty mod field. Avoid inserting
+		// these into the map to prevent garbage.
+		if s.mod == "" {
 			continue
 		}
+
+		mods[s.name] = s.mod
 	}
-	if scanner.Err() != nil {
-		return nil, scanner.Err()
+	if err := r.Err(); err != nil {
+		return nil, fmt.Errorf("reading kallsyms: %w", err)
 	}
+
 	return mods, nil
 }
 
-func LoadSymbolAddresses(symbols map[string]uint64) error {
+// AssignAddresses looks up the addresses of the requested symbols in the kernel
+// and assigns them to their corresponding values in the symbols map.
+//
+// Any symbols missing in the kernel are ignored. Returns an error if multiple
+// addresses were found for a symbol.
+func AssignAddresses(symbols map[string]uint64) error {
 	if len(symbols) == 0 {
 		return nil
 	}
@@ -87,41 +103,83 @@ func LoadSymbolAddresses(symbols map[string]uint64) error {
 		return err
 	}
 
-	if err := loadSymbolAddresses(f, symbols); err != nil {
+	if err := assignAddresses(f, symbols); err != nil {
 		return fmt.Errorf("error loading symbol addresses: %w", err)
 	}
 
 	return nil
 }
 
-func loadSymbolAddresses(f io.Reader, symbols map[string]uint64) error {
-	scan := bufio.NewScanner(f)
-	for scan.Scan() {
-		var (
-			addr   uint64
-			t      rune
-			symbol string
-		)
-
-		line := scan.Text()
-
-		_, err := fmt.Sscanf(line, "%x %c %s", &addr, &t, &symbol)
+// assignAddresses assigns kernel symbol addresses read from f to values
+// requested by symbols. Don't return when all symbols have been assigned, we
+// need to scan the whole thing to make sure the user didn't request an
+// ambiguous symbol.
+func assignAddresses(f io.Reader, symbols map[string]uint64) error {
+	r := newReader(f)
+	for r.Line() {
+		s, err, skip := parseSymbol(r, nil)
 		if err != nil {
-			return err
+			return fmt.Errorf("parsing kallsyms line: %w", err)
 		}
-		// Multiple addresses for a symbol have been found. Lets return an error to not confuse any
-		// users and handle it the same as libbpf.
-		if existingAddr, found := symbols[symbol]; existingAddr != 0 {
-			return fmt.Errorf("symbol %s(0x%x): duplicate found at address 0x%x %w",
-				symbol, existingAddr, addr, errKsymIsAmbiguous)
-		} else if found {
-			symbols[symbol] = addr
+		if skip {
+			continue
+		}
+
+		existing, requested := symbols[s.name]
+		if existing != 0 {
+			// Multiple addresses for a symbol have been found. Return a friendly
+			// error to avoid silently attaching to the wrong symbol. libbpf also
+			// rejects referring to ambiguous symbols.
+			return fmt.Errorf("symbol %s(0x%x): duplicate found at address 0x%x: %w", s.name, existing, s.addr, errAmbiguousKsym)
+		}
+		if requested {
+			symbols[s.name] = s.addr
 		}
 	}
-
-	if scan.Err() != nil {
-		return scan.Err()
+	if err := r.Err(); err != nil {
+		return fmt.Errorf("reading kallsyms: %w", err)
 	}
 
 	return nil
+}
+
+type ksym struct {
+	addr uint64
+	name string
+	mod  string
+}
+
+// parseSymbol parses a line from /proc/kallsyms into an address, type, name and
+// module. Skip will be true if the symbol doesn't match any of the given symbol
+// types. See `man 1 nm` for all available types.
+//
+// Example line: `ffffffffc1682010 T nf_nat_init  [nf_nat]`
+func parseSymbol(r *reader, types []rune) (s ksym, err error, skip bool) {
+	for i := 0; r.Word(); i++ {
+		switch i {
+		// Address of the symbol.
+		case 0:
+			s.addr, err = strconv.ParseUint(r.Text(), 16, 64)
+			if err != nil {
+				return s, fmt.Errorf("parsing address: %w", err), false
+			}
+		// Type of the symbol. Assume the character is ASCII-encoded by converting
+		// it directly to a rune, since it's a fixed field controlled by the kernel.
+		case 1:
+			if len(types) > 0 && !slices.Contains(types, rune(r.Bytes()[0])) {
+				return s, nil, true
+			}
+		// Name of the symbol.
+		case 2:
+			s.name = r.Text()
+		// Kernel module the symbol is provided by.
+		case 3:
+			s.mod = strings.Trim(r.Text(), "[]")
+		// Ignore any future fields.
+		default:
+			break
+		}
+	}
+
+	return
 }