@alban 2025-01-20
Hi @ti-mo,
Do you think the kallsyms package could be moved outside of "internal"?
I'm checking if the reader can be reused for some parts in [github.com/inspektor-gadget/inspektor-gadget/pkg/kallsyms](https://github.com/inspektor-gadget/inspektor-gadget/blob/main/pkg/kallsyms/kallsyms.go).