@ti-mo 2025-02-17
@alban I'm considering it, I'd like to use it for another project that I currently have no time to work on. Feel free to copy it in the meantime, you'll hear about it if we ever end up exporting it.